Refactor existing search, part 2
