+++ gramps_webapi/api/resources/base.py
@@ -38,12 +38,11 @@ from webargs import fields, validate
 from ...auth.const import PERM_ADD_OBJ, PERM_DEL_OBJ, PERM_EDIT_OBJ
 from ...const import GRAMPS_OBJECT_PLURAL
 from ..auth import require_permissions
-from ..search import SearchIndexer
+from ..search import SearchIndexer, get_search_indexer
 from ..util import (
     check_quota_people,
     get_db_handle,
     get_locale_for_language,
-    get_search_indexer,
     get_tree_from_jwt,
     update_usage_people,
     use_args,
