+++ gramps_webapi/api/resources/transactions.py
@@ -33,12 +33,11 @@ from webargs import fields
 
 from ...auth.const import PERM_ADD_OBJ, PERM_DEL_OBJ, PERM_EDIT_OBJ
 from ..auth import require_permissions
-from ..search import SearchIndexer
+from ..search import SearchIndexer, get_search_indexer
 from ..util import (
     abort_with_message,
     check_quota_people,
     get_db_handle,
-    get_search_indexer,
     get_tree_from_jwt,
     update_usage_people,
     use_args,
